Make admin-menu a Vue component
