+++ resources/views/master.blade.php
@@ -135,12 +135,6 @@ if (starts_with(app('VoyagerAuth')->user()->avatar, 'http://') || starts_with(ap
     @endif
 </script>
 @include('voyager::media.manager')
-@include('voyager::menu.admin_menu')
-<script>
-new Vue({
-    el: '#adminmenu',
-});
-</script>
 @yield('javascript')
 @stack('javascript')
 @if(!empty(config('voyager.additional_js')))<!-- Additional Javascript -->
